Removes hand coded intrinsics in favour for automatic vectorisation using clang
This code updates the viterbri27 to rely on clang for auto vectorisation and loop unrolling.
This means that the code will produce SIMD instructions on other architectures (for instance apple silicone).

The build system has been updated to remove the non portable code
And viterbri27 have been changed to ensure memory alignement and loop unrolling.